requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==4.2.13
django-environ==0.11.2
django-import-export==4.0.5
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.0
